+++ aeron-client/src/main/java/uk/co/real_logic/aeron/logbuffer/TermAppender.java
@@ -132,12 +132,12 @@ public class TermAppender
     public long claim(final HeaderWriter header, final int length, final BufferClaim bufferClaim)
     {
         final int frameLength = length + HEADER_LENGTH;
-        final int alignedLength = align(frameLength, FRAME_ALIGNMENT);
+        final int alignedLength = (int) align(frameLength, FRAME_ALIGNMENT);
         final long rawTail = getAndAddRawTail(alignedLength);
         final long termOffset = rawTail & 0xFFFF_FFFFL;
 
         final UnsafeBuffer termBuffer = this.termBuffer;
-        final int termLength = termBuffer.capacity();
+        final int termLength = (int) termBuffer.capacity();
 
         long resultingOffset = termOffset + alignedLength;
         if (resultingOffset > (termLength - HEADER_LENGTH))
@@ -168,12 +168,12 @@ public class TermAppender
         final HeaderWriter header, final DirectBuffer srcBuffer, final int srcOffset, final int length)
     {
         final int frameLength = length + HEADER_LENGTH;
-        final int alignedLength = align(frameLength, FRAME_ALIGNMENT);
+        final int alignedLength = (int) align(frameLength, FRAME_ALIGNMENT);
         final long rawTail = getAndAddRawTail(alignedLength);
         final long termOffset = rawTail & 0xFFFF_FFFFL;
 
         final UnsafeBuffer termBuffer = this.termBuffer;
-        final int termLength = termBuffer.capacity();
+        final int termLength = (int) termBuffer.capacity();
 
         long resultingOffset = termOffset + alignedLength;
         if (resultingOffset > (termLength - HEADER_LENGTH))
@@ -212,14 +212,16 @@ public class TermAppender
     {
         final int numMaxPayloads = length / maxPayloadLength;
         final int remainingPayload = length % maxPayloadLength;
-        final int lastFrameLength = remainingPayload > 0 ? align(remainingPayload + HEADER_LENGTH, FRAME_ALIGNMENT) : 0;
+        final int lastFrameLength = remainingPayload > 0
+                                  ? (int) align(remainingPayload + HEADER_LENGTH, FRAME_ALIGNMENT)
+                                  : 0;
         final int requiredLength = (numMaxPayloads * (maxPayloadLength + HEADER_LENGTH)) + lastFrameLength;
         final long rawTail = getAndAddRawTail(requiredLength);
         final int termId = termId(rawTail);
         final long termOffset = rawTail & 0xFFFF_FFFFL;
 
         final UnsafeBuffer termBuffer = this.termBuffer;
-        final int termLength = termBuffer.capacity();
+        final int termLength = (int) termBuffer.capacity();
 
         long resultingOffset = termOffset + requiredLength;
         if (resultingOffset > (termLength - HEADER_LENGTH))
@@ -235,7 +237,7 @@ public class TermAppender
             {
                 final int bytesToWrite = Math.min(remaining, maxPayloadLength);
                 final int frameLength = bytesToWrite + HEADER_LENGTH;
-                final int alignedLength = align(frameLength, FRAME_ALIGNMENT);
+                final int alignedLength = (int) align(frameLength, FRAME_ALIGNMENT);
 
                 header.write(termBuffer, offset, frameLength, termId);
                 termBuffer.putBytes(
