+++ aeron-client/src/test/java/uk/co/real_logic/aeron/logbuffer/TermReaderTest.java
@@ -44,7 +44,7 @@ public class TermReaderTest
     @Before
     public void setUp()
     {
-        when(termBuffer.capacity()).thenReturn(TERM_BUFFER_CAPACITY);
+        when(termBuffer.capacity()).thenReturn((long) TERM_BUFFER_CAPACITY);
     }
 
     @Test
@@ -114,7 +114,7 @@ public class TermReaderTest
     {
         final int msgLength = 1;
         final int frameLength = HEADER_LENGTH + msgLength;
-        final int alignedFrameLength = align(frameLength, FRAME_ALIGNMENT);
+        final int alignedFrameLength = (int) align(frameLength, FRAME_ALIGNMENT);
         final int termOffset = 0;
 
         when(termBuffer.getIntVolatile(0)).thenReturn(frameLength);
@@ -140,7 +140,7 @@ public class TermReaderTest
     {
         final int msgLength = 1;
         final int frameLength = HEADER_LENGTH + msgLength;
-        final int alignedFrameLength = align(frameLength, FRAME_ALIGNMENT);
+        final int alignedFrameLength = (int) align(frameLength, FRAME_ALIGNMENT);
         final int frameOffset = TERM_BUFFER_CAPACITY - alignedFrameLength;
 
         when(termBuffer.getIntVolatile(frameOffset)).thenReturn(frameLength);
@@ -160,7 +160,7 @@ public class TermReaderTest
     {
         final int msgLength = 1;
         final int frameLength = HEADER_LENGTH + msgLength;
-        final int alignedFrameLength = align(frameLength, FRAME_ALIGNMENT);
+        final int alignedFrameLength = (int) align(frameLength, FRAME_ALIGNMENT);
         final int frameOffset = TERM_BUFFER_CAPACITY - alignedFrameLength;
 
         when(termBuffer.getIntVolatile(frameOffset)).thenReturn(frameLength);
