+++ aeron-driver/src/main/java/uk/co/real_logic/aeron/driver/buffer/MappedRawLog.java
@@ -135,7 +135,7 @@ class MappedRawLog implements RawLog
     public ByteBuffer[] sliceTerms()
     {
         final ByteBuffer[] terms = new ByteBuffer[PARTITION_COUNT];
-        final int termLength = partitions[0].termBuffer().capacity();
+        final int termLength = (int) partitions[0].termBuffer().capacity();
 
         if (termLength < ONE_GIG)
         {
